+++ src/analysis/mdpp_sample_decoder_monitor_widget.cc
@@ -1,127 +1,72 @@
 #include "mdpp_sample_decoder_monitor_widget.h"
 
+#include <QStatusBar>
 #include <QTimer>
+#include <QtConcurrent>
+#include <mesytec-mvlc/util/stopwatch.h>
 
 #include "analysis/analysis.h"
-#include "util/qt_font.h"
 #include "ui_mdpp_sample_decoder_monitor_widget.h"
+#include "util/qt_font.h"
 
 using namespace mesytec::mvme;
 
 namespace analysis::ui
 {
 
-struct MdppSampleDecoderMonitorWidget::Private
+namespace
 {
-    MdppSampleDecoderMonitorWidget *q;
-    std::unique_ptr<Ui::MdppSampleDecoderMonitorWidget> ui;
-    std::shared_ptr<analysis::DataSourceMdppSampleDecoder> source_;
-    AnalysisServiceProvider *asp_;
-    QTimer refreshTimer_;
 
-    void refresh();
-};
-
-MdppSampleDecoderMonitorWidget::MdppSampleDecoderMonitorWidget(
-    const std::shared_ptr<analysis::DataSourceMdppSampleDecoder> &source,
-    AnalysisServiceProvider *serviceProvider,
-    QWidget *parent)
-: QMainWindow(parent)
-, d(std::make_unique<Private>())
-{
-    d->q = this;
-    d->ui = std::make_unique<Ui::MdppSampleDecoderMonitorWidget>();
-    d->ui->setupUi(this);
-    d->source_ = source;
-    d->asp_ = serviceProvider;
-
-    for (auto tb: { d->ui->tb_input, d->ui->tb_output, d->ui->tb_log })
-    {
-        tb->setReadOnly(true);
-        tb->setFont(make_monospace_font());
-    }
-
-    connect(d->ui->pb_pauseResumeRefresh, &QPushButton::clicked, this, [this]
-    {
-        if (d->ui->pb_pauseResumeRefresh->isChecked())
-            d->ui->pb_pauseResumeRefresh->setText("Resume");
-        else
-            d->ui->pb_pauseResumeRefresh->setText("Pause");
-    });
-
-    connect(d->ui->pb_refreshNow, &QPushButton::clicked, this, [this] { d->refresh(); });
-
-    connect(&d->refreshTimer_, &QTimer::timeout, this, [this]
-    {
-        if (!d->ui->pb_pauseResumeRefresh->isChecked())
-            d->refresh();
-    });
-
-    const auto refreshIntervals = { 0, 100, 500, 1000 };
-
-    for (const auto &interval: refreshIntervals)
-    {
-        d->ui->combo_refreshInterval->addItem(
-            interval == 0 ? "off" : QString::number(interval) + " ms",
-            interval);
-    }
-
-    connect(d->ui->combo_refreshInterval, QOverload<int>::of(&QComboBox::currentIndexChanged),
-            this, [this] (int index)
-    {
-        const auto interval = d->ui->combo_refreshInterval->itemData(index).toInt();
-        d->refreshTimer_.setInterval(interval);
-        if (interval == 0)
-            d->refreshTimer_.stop();
-        else
-            d->refreshTimer_.start();
-    });
-
-    d->ui->combo_refreshInterval->setCurrentIndex(1);
-}
-
-MdppSampleDecoderMonitorWidget::~MdppSampleDecoderMonitorWidget() = default;
-
-void MdppSampleDecoderMonitorWidget::Private::refresh()
+struct DecoderDebugResult
 {
     bool errorOrWarningSeen = false;
+    QString logText;
+    QString inputText;
+    QString outputText;
 
-    auto logger = [this, &errorOrWarningSeen] (const std::string &level, const std::string &message)
+    mesytec::mvlc::util::Stopwatch::duration_type dtDecode;
+};
+
+DecoderDebugResult run_decoder_collect_debug_info(
+    const mesytec::mvme::mdpp_sampling::DecodedMdppSampleEvent &decodedByAnalysis)
+{
+    mesytec::mvlc::util::Stopwatch sw;
+    DecoderDebugResult result;
+    QTextStream logStream(&result.logText);
+    QTextStream inputStream(&result.inputText);
+    QTextStream outputStream(&result.outputText);
+
+    auto logger = [&](const std::string &level, const std::string &message)
     {
-        ui->tb_log->append(fmt::format("[{}] {}", level, message).c_str());
+        logStream << fmt::format("[{}] {}\n", level, message).c_str();
 
         if (level == "warn" || level == "error")
         {
-            errorOrWarningSeen = true;
+            result.errorOrWarningSeen = true;
         }
     };
 
-    ui->tb_input->clear();
-    ui->tb_output->clear();
-    ui->tb_log->clear();
-
-    const auto decodedByAnalysis = source_->getDecodedEvent();
     const auto &inputData = decodedByAnalysis.inputData;
 
     if (decodedByAnalysis.moduleType.empty())
     {
-        ui->tb_output->append("module_type not set in analysis data. DAQ not running?");
-        return;
+        logStream << "module_type not set in analysis data. DAQ not running?\n";
+        return result;
     }
 
     // Rerun the decoder, redirecting log output and checking for errors and warnings.
-    auto decoded = mdpp_sampling::decode_mdpp_samples(
-        inputData.data(), inputData.size(), decodedByAnalysis.moduleType.c_str(), logger);
+    auto decoded = mdpp_sampling::decode_mdpp_samples(inputData.data(), inputData.size(),
+                                                      decodedByAnalysis.moduleType.c_str(), logger);
 
-    for (size_t i=0; i<inputData.size(); ++i)
+    for (size_t i = 0; i < inputData.size(); ++i)
     {
-        ui->tb_input->append(fmt::format("{:3d}: {:#010x}", i, inputData[i]).c_str());
+        inputStream << fmt::format("{:3d}: {:#010x}\n", i, inputData[i]).c_str();
     }
 
     std::ostringstream oss;
     oss << fmt::format("module_type={}, input size={}\n", decoded.moduleType, inputData.size());
-    oss << fmt::format("module_header={:#010x}, moduleId={:04x}, timestamp={}\n",
-        decoded.header, decoded.headerModuleId, decoded.timestamp);
+    oss << fmt::format("module_header={:#010x}, moduleId={:04x}, timestamp={}\n", decoded.header,
+                       decoded.headerModuleId, decoded.timestamp);
     oss << fmt::format("#traces={}\n", decoded.traces.size());
 
     if (!decoded.traces.empty())
@@ -129,7 +74,7 @@ void MdppSampleDecoderMonitorWidget::Private::refresh()
         oss << "\n";
         const auto &traces = decoded.traces;
 
-        for (size_t traceIndex=0; traceIndex<static_cast<size_t>(traces.size()); ++traceIndex)
+        for (size_t traceIndex = 0; traceIndex < static_cast<size_t>(traces.size()); ++traceIndex)
         {
             auto thDebug = traces[traceIndex].traceHeader.parts.debug;
             auto thConfig = traces[traceIndex].traceHeader.parts.config;
@@ -138,23 +83,159 @@ void MdppSampleDecoderMonitorWidget::Private::refresh()
 
             oss << "trace #" << traceIndex << ":\n";
 
-            oss <<  fmt::format("  channel={}\n", traces[traceIndex].channel);
+            oss << fmt::format("  channel={}\n", traces[traceIndex].channel);
 
-            oss <<  fmt::format("  channel={}, #samples={}, trace header={:#010x}, .debug={}, .config={}, .phase={}, .length={}\n",
-                traces[traceIndex].channel, traces[traceIndex].samples.size(),
-                traces[traceIndex].traceHeader.value, thDebug, thConfig, thPhase, thLength);
+            oss << fmt::format("  channel={}, #samples={}, trace header={:#010x}, .debug={}, "
+                               ".config={}, .phase={}, .length={}\n",
+                               traces[traceIndex].channel, traces[traceIndex].samples.size(),
+                               traces[traceIndex].traceHeader.value, thDebug, thConfig, thPhase,
+                               thLength);
 
-            oss <<  fmt::format("  channel={}, samples: {}\n",
-                traces[traceIndex].channel, fmt::join(traces[traceIndex].samples, ", "));
+            oss << fmt::format("  channel={}, samples: {}\n", traces[traceIndex].channel,
+                               fmt::join(traces[traceIndex].samples, ", "));
         }
     }
 
-    ui->tb_output->append(oss.str().c_str());
+    outputStream << oss.str().c_str();
+    result.dtDecode = sw.interval();
 
-    if (errorOrWarningSeen && ui->cb_holdRefreshOnError->isChecked())
+    return result;
+}
+
+} // namespace
+
+struct MdppSampleDecoderMonitorWidget::Private
+{
+    MdppSampleDecoderMonitorWidget *q;
+    std::unique_ptr<Ui::MdppSampleDecoderMonitorWidget> ui;
+    std::shared_ptr<analysis::DataSourceMdppSampleDecoder> source_;
+    AnalysisServiceProvider *asp_;
+    QTimer refreshTimer_;
+    QFutureWatcher<DecoderDebugResult> debugResultWatcher_;
+
+    void startRefresh();
+    void onRefreshDone();
+};
+
+MdppSampleDecoderMonitorWidget::MdppSampleDecoderMonitorWidget(
+    const std::shared_ptr<analysis::DataSourceMdppSampleDecoder> &source,
+    AnalysisServiceProvider *serviceProvider, QWidget *parent)
+    : QMainWindow(parent)
+    , d(std::make_unique<Private>())
+{
+    d->q = this;
+    d->ui = std::make_unique<Ui::MdppSampleDecoderMonitorWidget>();
+    d->ui->setupUi(this);
+    d->source_ = source;
+    d->asp_ = serviceProvider;
+    d->refreshTimer_.setSingleShot(true);
+
+    for (auto tb: {d->ui->tb_input, d->ui->tb_output, d->ui->tb_log})
+    {
+        tb->setReadOnly(true);
+        tb->setFont(make_monospace_font());
+    }
+
+    connect(d->ui->pb_pauseResumeRefresh, &QPushButton::clicked, this,
+            [this](bool checked)
+            {
+                if (checked)
+                {
+                    d->ui->pb_pauseResumeRefresh->setText("Resume");
+                }
+                else
+                {
+                    d->ui->pb_pauseResumeRefresh->setText("Pause");
+                    d->refreshTimer_.start();
+                }
+            });
+
+    connect(d->ui->pb_refreshNow, &QPushButton::clicked, this, [this] { d->startRefresh(); });
+
+    connect(&d->refreshTimer_, &QTimer::timeout, this,
+            [this]
+            {
+                if (!d->ui->pb_pauseResumeRefresh->isChecked())
+                    d->startRefresh();
+            });
+
+    const auto refreshIntervals = {0, 100, 500, 1000};
+
+    for (const auto &interval: refreshIntervals)
+    {
+        d->ui->combo_refreshInterval->addItem(
+            interval == 0 ? "off" : QString::number(interval) + " ms", interval);
+    }
+
+    connect(d->ui->combo_refreshInterval, QOverload<int>::of(&QComboBox::currentIndexChanged), this,
+            [this](int index)
+            {
+                const auto interval = d->ui->combo_refreshInterval->itemData(index).toInt();
+                d->refreshTimer_.setInterval(interval);
+                if (interval == 0)
+                    d->refreshTimer_.stop();
+                else
+                    d->refreshTimer_.start();
+            });
+
+    d->ui->combo_refreshInterval->setCurrentIndex(1);
+
+    connect(&d->debugResultWatcher_, &QFutureWatcher<DecoderDebugResult>::finished, this,
+            [this] { d->onRefreshDone(); });
+}
+
+MdppSampleDecoderMonitorWidget::~MdppSampleDecoderMonitorWidget()
+{
+    d->refreshTimer_.stop();
+    d->debugResultWatcher_.waitForFinished();
+}
+
+void MdppSampleDecoderMonitorWidget::Private::startRefresh()
+{
+    if (debugResultWatcher_.isRunning())
+    {
+        qDebug() << __PRETTY_FUNCTION__ << "early return, previous refresh still in progress";
+        return;
+    }
+
+    auto future = QtConcurrent::run(run_decoder_collect_debug_info, source_->getDecodedEvent());
+    debugResultWatcher_.setFuture(future);
+}
+
+void MdppSampleDecoderMonitorWidget::Private::onRefreshDone()
+{
+    mesytec::mvlc::util::Stopwatch sw;
+    auto result = debugResultWatcher_.result();
+
+    if (ui->tb_input->toPlainText() != result.inputText)
+        ui->tb_input->setText(result.inputText);
+
+    if (ui->tb_output->toPlainText() != result.outputText)
+        ui->tb_output->setText(result.outputText);
+
+    if (ui->tb_log->toPlainText() != result.logText)
+        ui->tb_log->setText(result.logText);
+
+    if (result.errorOrWarningSeen && ui->cb_holdRefreshOnError->isChecked())
     {
         ui->pb_pauseResumeRefresh->setChecked(true);
     }
+
+    // pull the current selected refresh interval from the ui and kick of the timer again.
+    const auto interval = ui->combo_refreshInterval->currentData().toInt();
+    refreshTimer_.setInterval(interval);
+    refreshTimer_.setSingleShot(true);
+    if (interval == 0)
+        refreshTimer_.stop();
+    else
+        refreshTimer_.start();
+
+    auto dtRefresh = sw.interval();
+
+    q->statusBar()->showMessage(
+        QString("Decode time: %1 ms, UI update time: %2 ms")
+            .arg(result.dtDecode.count() / 1000.0, 0, 'f', 1)
+            .arg(dtRefresh.count() / 1000.0, 0, 'f', 1));
 }
 
-}
+} // namespace analysis::ui